Add Parser and TaskList classes
