+++ src/main/java/TaskType.java
@@ -1,0 +1,5 @@
+enum TaskType {
+    TODO,
+    EVENT,
+    DEADLINE
+}
